+++ bank-feeds/src/sdk/models/shared/index.ts
@@ -34,3 +34,4 @@ export * from "./sourceaccount";
 export * from "./targetaccountoption";
 export * from "./validation";
 export * from "./validationitem";
+export * from "./zero";